+++ projects/forty-cdk/carousel/src/public-api.ts
@@ -1,4 +1,4 @@
-export { ForCarousel, provideForCarousel } from './carousel';
+export { ForCarousel } from './carousel';
 export { ForCarouselViewport } from './carousel-viewport';
 export { ForCarouselTrack } from './carousel-track';
 export { ForCarouselSlide } from './carousel-slide';